requirements.txt
openai==1.43.0
streamlit==1.38.0
python-dotenv==1.0.1